# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test task API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/business-activity": {
            "get": {
                "tags": [
                    "BusinessActivity"
                ],
                "description": "Get full tree of activities",
                "operationId": "fe5f4ead8993f0059716cd4d82709016",
                "responses": {
                    "200": {
                        "description": "Tree of activities",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BusinessActivityTreeItem"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "http": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "BusinessActivity"
                ],
                "description": "Create an activity",
                "operationId": "f658d1cdc274832bf291bf1a150d3ad0",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateBusinessActivityRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Activity created"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Cannot find parent activity by id"
                    },
                    "409": {
                        "description": "Id already taken|Max tree depth exceeded"
                    }
                },
                "security": [
                    {
                        "http": []
                    }
                ]
            }
        },
...
        "/api/business-activity/{id}": {
            "get": {
                "tags": [
                    "BusinessActivity"
                ],
                "description": "Get activity subtree",
                "operationId": "a9927395040ae17626e3f21ad4610ebf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tree of activities",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BusinessActivityTreeItem"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cannot find activity by id"
                    }
                },
                "security": [
                    {
                        "http": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "BusinessActivity"
                ],
                "description": "Update activity",
                "operationId": "8ef8cc93f2b5c93c85129edad43895b8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateBusinessActivityRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Activity updated"
                    },
                    "404": {
                        "description": "Cannot find activity by id"
                    }
                },
                "security": [
                    {
                        "http": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "BusinessActivity"
                ],
                "description": "Delete a company",
                "operationId": "b47807edc2e43b87e108c580103eed4f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Activity deleted"
                    },
                    "404": {
                        "description": "Cannot find activity by id"
                    }
                },
                "security": [
                    {
                        "http": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "BusinessActivity"
                ],
                "description": "Move activitity within a tree",
                "operationId": "d09e145b447a50fbb3ac7a3b907061ab",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MoveBusinessActivityRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "New activity tree",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BusinessActivityTreeItem"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cannot find activity by id"
                    },
                    "409": {
                        "description": "Max tree depth exceeded"
                    }
                },
                "security": [
                    {
                        "http": []
                    }
                ]
            }
        },
        "/api/business-activity/{id}/companies": {
            "get": {
                "tags": [
                    "BusinessActivity"
                ],
                "description": "Get companies by activity",
                "operationId": "c7c34fb560aaadbe06aa275c4da6303d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of companies",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CompanyDTO"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cannot find activity by id"
                    }
                },
                "security": [
                    {
                        "http": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CompanyDTO": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "buildingId": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "CreateBusinessActivityRequest": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "parentId": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "MoveBusinessActivityRequest": {
                "properties": {
                    "parentId": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UpdateBusinessActivityRequest": {
                "properties": {
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "BusinessActivityTreeItem": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "childs": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/BusinessActivityTreeItem"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "http": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}